merge

+++ code/modules/modular_computers/computers/item/computer_damage.dm
@@ -18,11 +18,13 @@
 
 /obj/item/modular_computer/proc/break_apart()
 	if(!(flags_1 & NODECONSTRUCT_1))
-		physical.visible_message("\The [src] breaks apart!")
+		physical.visible_message("<span class='notice'>\The [src] breaks apart!</span>")
 		var/turf/newloc = get_turf(src)
 		new /obj/item/stack/sheet/metal(newloc, round(steel_sheet_cost/2))
 		for(var/C in all_components)
 			var/obj/item/computer_hardware/H = all_components[C]
+			if(QDELETED(H))
+				continue
 			uninstall_component(H)
 			H.forceMove(newloc)
 			if(prob(25))
